requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic_core==2.14.1